Move `EXPERIMENTAL_setMessagePathDropConfig` to internal types
