+++ packages/studio-base/src/Workspace.stories.tsx
@@ -6,10 +6,10 @@ import { StoryObj } from "@storybook/react";
 import { fireEvent, screen, waitFor } from "@storybook/testing-library";
 import { useEffect, useState } from "react";
 
-import { DraggedMessagePath } from "@foxglove/studio";
 import MultiProvider from "@foxglove/studio-base/components/MultiProvider";
 import Panel from "@foxglove/studio-base/components/Panel";
 import { usePanelContext } from "@foxglove/studio-base/components/PanelContext";
+import { DraggedMessagePath } from "@foxglove/studio-base/components/PanelExtensionAdapter";
 import PanelToolbar from "@foxglove/studio-base/components/PanelToolbar";
 import { LayoutData } from "@foxglove/studio-base/context/CurrentLayoutContext";
 import PanelCatalogContext, {
