+++ packages/studio-base/src/components/PanelContext.ts
@@ -11,7 +11,7 @@
 //   found at http://www.apache.org/licenses/LICENSE-2.0
 //   You may not use this file except in compliance with the License.
 
-import { MessagePathDropConfig } from "@foxglove/studio";
+import { MessagePathDropConfig } from "@foxglove/studio-base/components/PanelExtensionAdapter";
 import { SaveConfig, PanelConfig, OpenSiblingPanel } from "@foxglove/studio-base/types/panels";
 
 export type PanelContextType<T> = {
