+++ packages/studio-base/src/components/PanelExtensionAdapter/PanelExtensionAdapter.tsx
@@ -508,7 +508,7 @@ function PanelExtensionAdapter(
         setDefaultPanelTitle(title);
       },
 
-      EXPERIMENTAL_setMessagePathDropConfig(dropConfig) {
+      unstable_setMessagePathDropConfig(dropConfig) {
         setMessagePathDropConfig(dropConfig);
       },
     };
